test/misc: done
