+++ test/misc/Makefile
@@ -1,49 +1,48 @@
 #--
-TOP = $(abspath $(dir $(firstword $(MAKEFILE_LIST)))../../)
-TEST = $(TOP)/test
-SRC = $(TOP)/src
-
-#--
 CC      ?= clang
 CFLAGS  ?= -O3 -Wall -Wextra -Wpedantic -Wvla -Werror -std=c99 \
 	         -Wundef -Wshadow -Wcast-align -Wpointer-arith -Wmissing-prototypes \
 	         -fstrict-aliasing -fno-common -pipe
 
-JASMIN ?= jasminc
+#--
+TOP  = $(abspath $(dir $(firstword $(MAKEFILE_LIST)))../../)
+TEST = $(TOP)/test
+SRC  = $(TOP)/src
+
+JASMIN   ?= jasminc
 JINCLUDE ?= -I Libjbn:$(SRC)/
 
-BUILD ?= $(TEST)/bin/globals
-
+PARAMS_DIR ?= $(TEST)/bin/params
 MIN_NLIMBS ?= 1
 MAX_NLIMBS ?= 16
 MAX_TESTS  ?= 10
 
 #--
-default: globals
+default: $(PARAMS_DIR)
 
 #--
-gen: gen.c
-	$(CC) $(CFLAGS) -D_MAIN_GEN -o $@ $< -lgmp
+params: params.c
+	$(CC) $(CFLAGS) -D_MAIN_PARAMS -o $@ $< -lgmp
 
-primes-det: primes.c gen.c ../common/notrandombytes.c
+primes-det: primes.c params.c ../common/notrandombytes.c
 	$(CC) $(CFLAGS) -D_MAIN_PRIMES -o $@ -I../common $^ -lgmp
 
-primes-rnd: primes.c gen.c ../common/randombytes.c
+primes-rnd: primes.c params.c ../common/randombytes.c
 	$(CC) $(CFLAGS) -D_MAIN_PRIMES -o $@ -I../common $^ -lgmp
 
 #--
-JAZZ = $(shell find $(BUILD) -name '*.jazz')
-ASM  = $(JAZZ:%.jazz=%.s)
-
-.PHONY: globals
-globals: gen primes-det
-	./sike $(BUILD)
-	./det $(BUILD) $(MIN_NLIMBS) $(MAX_NLIMBS) $(MAX_TESTS)
-	$(MAKE) asm
+.PHONY: $(PARAMS_DIR)
+$(PARAMS_DIR): params primes-det
+	./sike.sh $(PARAMS_DIR)
+	./gen.sh det $(PARAMS_DIR) $(MIN_NLIMBS) $(MAX_NLIMBS) $(MAX_TESTS)
+	$(MAKE) compile
 
 #--
-.PHONY: asm
-asm: $(ASM)
+.PHONY: compile
+compile: JAZZ := $(shell find $(PARAMS_DIR) -name '*.jazz')
+compile: ASM  := $(JAZZ:%.jazz=%.s)
+compile:
+	$(MAKE) $(ASM)
 
 %.s: %.jazz
 	$(JASMIN) -o $@ $(JINCLUDE) $<
@@ -51,7 +50,7 @@ asm: $(ASM)
 #--
 .PHONY: clean
 clean:
-	rm -f gen primes-det primes-rnd
+	rm -f params primes-det primes-rnd
 
 distclean: clean
-	rm -fr $(BUILD)
+	rm -fr $(PARAMS_DIR)
